fix(kadena-cli) fix test, and seperated  createTransactionWidthDetails for multiusage
