+++ packages/tools/kadena-cli/src/tx/tests/transaction-live.test.ts
@@ -1,12 +1,12 @@
-import { describe, it } from 'vitest'; // expect,
+import path from 'node:path';
+import { describe, expect, it } from 'vitest';
+import { ensureNetworksConfiguration } from '../../networks/utils/networkHelpers.js';
 import { services } from '../../services/index.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { defaultTemplates } from '../commands/templates/templates.js';
 import { createTransaction } from '../commands/txCreateTransaction.js';
-// import {
-//   createTransactionWithDetails,
-//   testTransactionAction,
-// } from '../commands/txTestSignedTransaction.js';
+import { testTransactionAction } from '../commands/txTestSignedTransaction.js';
+import { createTransactionWithDetails } from '../utils/txHelpers.js';
 import { signTransactionFileWithKeyPairAction } from '../utils/txSignWithKeypair.js';
 
 const publicKey =
@@ -19,6 +19,10 @@ const targetAccount =
 describe('template to live test', () => {
   // NOTE: this tests uses live testnet04 meaning it is not isolated!!!
   it('creates, signs and tests the transaction', async () => {
+    const root = path.join(__dirname, '../../../');
+    const kadenaDir = path.join(root, '.kadena');
+    await ensureNetworksConfiguration(kadenaDir);
+
     const variables = {
       'account:from': `k:${publicKey}`,
       'account:to': targetAccount,
@@ -42,21 +46,19 @@ describe('template to live test', () => {
     });
     assertCommandError(signed);
 
-    // console.dir(JSON.parse(signed.data.commasnds[0].cmd), { depth: Infinity });
-    // console.log(signed.data.commands[0]);
+    const transactionsWithDetails = await createTransactionWithDetails(
+      signed.data.commands.map((c) => c.command),
+      { txTransactionNetwork: ['testnet'] },
+    );
 
-    // const transactionsWithDetails = await createTransactionWithDetails(
-    //   signed.data.commands.map((c) => c.command),
-    //   { txTransactionNetwork: ['testnet'] },
-    // );
+    const test = await testTransactionAction({
+      transactionsWithDetails,
+    });
+    assertCommandError(test);
 
-    // const test = await testTransactionAction({
-    //   transactionsWithDetails,
-    // });
-    // assertCommandError(test);
-    // expect(test.data[0].result).toEqual({
-    //   status: 'success',
-    //   data: 'Write succeeded',
-    // });
+    expect(test.data.transactions[0].response?.result).toEqual({
+      status: 'success',
+      data: 'Write succeeded',
+    });
   });
 });
